models/engine/file_storage.py
#!/usr/bin/python3
"""
Module that defines a classthat serializes instances
to a JSON file and deserializes JSON file to instances
"""
import json


class FileStorage:
    """This class manages storage in JSON format"""
    __file_path = 'file.json'
    __objects = {}

    def all(self):
        """Returns a dictionary of objects currently in storage"""
        return FileStorage.__objects

    def new(self, obj):
        """Adds new object to storage dictionary"""
        self.all().update({obj.to_dict()['__class__'] + '.' + obj.id: obj})

    def save(self):
        """serializes __objects to the JSON file (path: __file_path)"""
        with open(FileStorage.__file_path, 'w') as f:
            tmp = {}
            tmp.update(FileStorage.__objects)
            for key, val in tmp.items():
                tmp[key] = val.to_dict()
            json.dump(tmp, f)

    def reload(self):
        """ deserializes the JSON file to __objects"""
        from models.base_model import BaseModel
        from models.user import User

        cls = {
                    'BaseModel': BaseModel, 'User': User
                  }
        try:
            tmp = {}
            with open(FileStorage.__file_path, 'r') as f:
                tmp = json.load(f)
                for key, val in tmp.items():
                    self.all()[key] = cls[val['__class__']](**val)
        except FileNotFoundError:
            pass
